Inlining _upsert_lis_result_sourceid into the view
This follows on from previous work and has the same separated unit
tests which may not be the way to go.
